# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rfc3339.date — Time API",
    "version": "1.0.0",
    "summary": "Strict RFC3339 time endpoints + validation + conversions (incl. exotic time standards)",
    "description": "A precise, developer-centric time API. Defaults to canonical RFC3339 UTC for /now, supports IANA time zones, strict RFC3339 validation, ISO8601 validation, and conversions across common + exotic representations (NTP, GPS, TAI, JD/MJD, Excel, HTTP-date, RFC5322). Plain text is first-class; JSON is optional via Accept header or ?json=1.\n",
    "license": {
      "name": "UNLICENSED",
      "identifier": "NOASSERTION"
    }
  },
  "servers": [
    {
      "url": "https://rfc3339.date",
      "description": "Production"
    },
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local dev (wrangler)"
    }
  ],
...
  "security": [],
  "paths": {
...
    "/canonical": {
      "get": {
        "operationId": "canonicalizeTimestamp",
        "tags": [
          "Conversion",
          "Validation"
        ],
        "summary": "Canonicalize a timestamp representation",
        "description": "Parses the given value under the selected profile and returns canonical RFC3339 UTC by default. If out is set, returns canonical output for that target format.\n",
        "parameters": [
          {
            "name": "value",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "profile",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "rfc3339",
                "iso8601",
                "iso8601:strict",
                "iso8601:extended",
                "iso8601:basic"
              ],
              "default": "rfc3339"
            },
            "description": "Validation profile."
          },
          {
            "name": "out",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "rfc3339",
                "iso8601",
                "unix",
                "unixms",
                "ntp",
                "httpdate",
                "emaildate",
                "gps",
                "tai",
                "jd",
                "mjd",
                "excel1900",
                "excel1904",
                "weekdate",
                "ordinal",
                "doy"
              ],
              "default": "rfc3339"
            },
            "description": "Output encoding."
          },
          {
            "name": "precision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9,
              "default": 3
            },
            "description": "Fractional second digits to emit (0..9)."
          },
          {
            "name": "tz",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "IANA time zone for rendering outputs that are zone-dependent."
          },
          {
            "name": "json",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ],
              "default": 0
            },
            "description": "If 1, forces JSON output regardless of Accept header."
          }
        ],
        "responses": {
          "200": {
            "description": "Canonical output",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "canon": {
                    "value": "2026-02-25T19:17:03Z"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "in",
                    "out",
                    "value_in",
                    "value_out"
                  ],
                  "properties": {
                    "in": {
                      "type": "string"
                    },
                    "out": {
                      "type": "string"
                    },
                    "value_in": {
                      "type": "string"
                    },
                    "value_out": {
                      "type": "string"
                    },
                    "tz": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "precision": {
                      "type": "integer"
                    },
                    "instant": {
                      "type": "object",
                      "description": "Canonical internal instant representation.",
                      "properties": {
                        "rfc3339z": {
                          "type": "string"
                        },
                        "unix": {
                          "type": "integer"
                        },
                        "unixms": {
                          "type": "integer"
                        }
                      }
                    },
                    "notes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "candidates": {
                      "type": "array",
                      "description": "Present when in=auto is ambiguous.",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid input, ambiguity, or unsupported combination)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found (unknown time zone or resource)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/convert": {
      "get": {
        "operationId": "convertTimestamp",
        "tags": [
          "Conversion"
        ],
        "summary": "Convert between time encodings",
        "description": "Converts an input value from an input encoding to an output encoding. If in=auto, ambiguity results in 400 with candidates. For timezone-dependent output formats, supply tz=IANA. For standards requiring leap-second tables (TAI/GPS), supply leapdata=latest or a known version.\n",
        "parameters": [
          {
            "name": "value",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "in",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "auto",
                "rfc3339",
                "iso8601",
                "unix",
                "unixms",
                "ntp",
                "httpdate",
                "emaildate",
                "gps",
                "tai",
                "jd",
                "mjd",
                "excel1900",
                "excel1904",
                "weekdate",
                "ordinal",
                "doy"
              ],
              "default": "auto"
            },
            "description": "Input encoding."
          },
          {
            "name": "out",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "rfc3339",
                "iso8601",
                "unix",
                "unixms",
                "ntp",
                "httpdate",
                "emaildate",
                "gps",
                "tai",
                "jd",
                "mjd",
                "excel1900",
                "excel1904",
                "weekdate",
                "ordinal",
                "doy"
              ],
              "default": "rfc3339"
            },
            "description": "Output encoding."
          },
          {
            "name": "precision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9,
              "default": 3
            },
            "description": "Fractional second digits to emit (0..9)."
          },
          {
            "name": "tz",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "IANA time zone for rendering outputs that are zone-dependent."
          },
          {
            "name": "leapdata",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "latest"
            },
            "description": "Leap-second dataset version identifier, or \"latest\"."
          },
          {
            "name": "json",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ],
              "default": 0
            },
            "description": "If 1, forces JSON output regardless of Accept header."
          }
        ],
        "responses": {
          "200": {
            "description": "Converted output",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "rfc3339": {
                    "value": "2023-11-14T22:13:20Z"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "in",
                    "out",
                    "value_in",
                    "value_out"
                  ],
                  "properties": {
                    "in": {
                      "type": "string"
                    },
                    "out": {
                      "type": "string"
                    },
                    "value_in": {
                      "type": "string"
                    },
                    "value_out": {
                      "type": "string"
                    },
                    "tz": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "precision": {
                      "type": "integer"
                    },
                    "instant": {
                      "type": "object",
                      "description": "Canonical internal instant representation.",
                      "properties": {
                        "rfc3339z": {
                          "type": "string"
                        },
                        "unix": {
                          "type": "integer"
                        },
                        "unixms": {
                          "type": "integer"
                        }
                      }
                    },
                    "notes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "candidates": {
                      "type": "array",
                      "description": "Present when in=auto is ambiguous.",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid input, ambiguity, or unsupported combination)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found (unknown time zone or resource)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/round": {
      "get": {
        "operationId": "roundTimestamp",
        "tags": [
          "Conversion"
        ],
        "summary": "Round a timestamp to a unit",
        "parameters": [
          {
            "name": "value",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "in",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "auto",
                "rfc3339",
                "iso8601",
                "unix",
                "unixms",
                "ntp",
                "httpdate",
                "emaildate",
                "gps",
                "tai",
                "jd",
                "mjd",
                "excel1900",
                "excel1904",
                "weekdate",
                "ordinal",
                "doy"
              ],
              "default": "auto"
            },
            "description": "Input encoding."
          },
          {
            "name": "unit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ns",
                "us",
                "ms",
                "s",
                "min",
                "hour",
                "day"
              ]
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "floor",
                "ceil",
                "nearest"
              ],
              "default": "nearest"
            }
          },
          {
            "name": "out",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "rfc3339",
                "iso8601",
                "unix",
                "unixms",
                "ntp",
                "httpdate",
                "emaildate",
                "gps",
                "tai",
                "jd",
                "mjd",
                "excel1900",
                "excel1904",
                "weekdate",
                "ordinal",
                "doy"
              ],
              "default": "rfc3339"
            },
            "description": "Output encoding."
          },
          {
            "name": "precision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9,
              "default": 3
            },
            "description": "Fractional second digits to emit (0..9)."
          },
          {
            "name": "tz",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "IANA time zone for rendering outputs that are zone-dependent."
          },
          {
            "name": "json",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ],
              "default": 0
            },
            "description": "If 1, forces JSON output regardless of Accept header."
          }
        ],
        "responses": {
          "200": {
            "description": "Rounded output",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "in",
                    "out",
                    "value_in",
                    "value_out"
                  ],
                  "properties": {
                    "in": {
                      "type": "string"
                    },
                    "out": {
                      "type": "string"
                    },
                    "value_in": {
                      "type": "string"
                    },
                    "value_out": {
                      "type": "string"
                    },
                    "tz": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "precision": {
                      "type": "integer"
                    },
                    "instant": {
                      "type": "object",
                      "description": "Canonical internal instant representation.",
                      "properties": {
                        "rfc3339z": {
                          "type": "string"
                        },
                        "unix": {
                          "type": "integer"
                        },
                        "unixms": {
                          "type": "integer"
                        }
                      }
                    },
                    "notes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "candidates": {
                      "type": "array",
                      "description": "Present when in=auto is ambiguous.",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid input, ambiguity, or unsupported combination)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}